+++ src/lua/builtins/mod.rs
@@ -11,6 +11,7 @@ mod process;
 mod serialization;
 mod shell;
 mod template;
+mod temporal;
 mod ws;
 
 pub fn register_all(lua: &mlua::Lua, client: reqwest::Client) -> mlua::Result<()> {
@@ -36,5 +37,6 @@ pub fn register_all(lua: &mlua::Lua, client: reqwest::Client) -> mlua::Result<()
     process::register_process(lua)?;
     disk::register_disk(lua)?;
     os_info::register_os(lua)?;
+    temporal::register_temporal(lua)?;
     Ok(())
 }